EQTL with batchtools

+++ SSH-CLUSTER/delete.py
@@ -10,6 +10,8 @@ res = client.terminate_instances(InstanceIds=cluster['iids'])
 
 for id in cluster['iids']:
   boto3.resource('ec2').Instance(id).wait_until_terminated()
-client.delete_security_group(GroupName=cluster['sgname'])
+
+if cluster['delete-sg']:
+  client.delete_security_group(GroupName=cluster['sgname'])
